+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Road map
 
+- [x] Option to not show when Error List is visible
 - [ ] Flash on save
 
 Features that have a checkmark are complete and available for